feat: support non-templated Discussions
We can assume that if a template cannot be found for a Discussion, we
can "just" treat it as if it's a body-only Discussion, which is how the
GitHub UI would present it.

If there's a different error trying to retrieve the Discussion template,
we can then return that error.

Closes #2.